cookie stats reworked - single file
